% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_download.R
\name{hub_download}
\alias{hub_download}
\title{Hub download}
\usage{
hub_download(
  repo_id,
  filename,
  ...,
  revision = "main",
  repo_type = "model",
  local_files_only = FALSE,
  force_download = FALSE
)
}
\description{
Download data from hf hub.
}
