% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_download.R
\name{WEIGHTS_NAME}
\alias{WEIGHTS_NAME}
\alias{WEIGHTS_INDEX_NAME}
\title{Constants related to filenames in the HUB}
\usage{
WEIGHTS_NAME()

WEIGHTS_INDEX_NAME()
}
\description{
Constants related to filenames in the HUB
}
\section{Functions}{
\itemize{
\item \code{WEIGHTS_NAME()}: Name of the weights file in HF hub.

\item \code{WEIGHTS_INDEX_NAME()}: Name of the weights index file in HF hub.

}}
