% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_info.R
\name{hub_repo_info}
\alias{hub_repo_info}
\alias{hub_dataset_info}
\title{Queries information about Hub repositories}
\usage{
hub_repo_info(
  repo_id,
  ...,
  repo_type = NULL,
  revision = NULL,
  files_metadata = FALSE
)

hub_dataset_info(repo_id, ..., revision = NULL, files_metadata = FALSE)
}
\arguments{
\item{repo_id}{The repository identifier, eg \code{"bert-base-uncased"} or \code{"deepset/sentence_bert"}.}

\item{...}{currenytly unused.}

\item{repo_type}{The type of the repository. Currently only \code{"model"} is supported.}

\item{revision}{Revision (branch, tag or commitid) to download the file from.}

\item{files_metadata}{Obtain files metadata information when querying repository information.}
}
\description{
Queries information about Hub repositories
}
\section{Functions}{
\itemize{
\item \code{hub_dataset_info()}: Query information from a Hub Dataset

}}
