% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_download.R
\name{WEIGHTS_NAME}
\alias{WEIGHTS_NAME}
\alias{WEIGHTS_INDEX_NAME}
\title{Weight file names in HUB}
\usage{
WEIGHTS_NAME()

WEIGHTS_INDEX_NAME()
}
\value{
A string with the default file names for indexes in the Hugging Face Hub.
}
\description{
Weight file names in HUB
}
\section{Functions}{
\itemize{
\item \code{WEIGHTS_NAME()}: Name of weights file

\item \code{WEIGHTS_INDEX_NAME()}: Name of weights index file

}}
\examples{
WEIGHTS_NAME()
WEIGHTS_INDEX_NAME()
}
